Filtering by job attributes works
